refactor: (#801) CommandNoticeService 로 ScheduleVoteResultNotice 메서드 옮김
