--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/service/VoteResultNoticeSchedulerService.kt
@@ -1,9 +1,0 @@
-package team.aliens.dms.domain.vote.service
-
-import java.time.LocalDateTime
-import java.util.*
-
-interface VoteResultNoticeSchedulerService {
-
-    fun scheduleVoteResultNotice(savedVotingTopicId: UUID, reservedTime: LocalDateTime)
-}
