+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/service/VoteService.kt
@@ -1,0 +1,7 @@
+package team.aliens.dms.domain.vote.service
+
+class VoteService(
+    val commendVotingTopicService: CommendVotingTopicService,
+    val getVotingTopicService: GetVotingTopicService,
+) : CommendVotingTopicService by commendVotingTopicService,
+    GetVotingTopicService by getVotingTopicService
