+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/usecase/CreateVotingTopicUseCase.kt
@@ -2,17 +2,17 @@ package team.aliens.dms.domain.vote.usecase
 
 import team.aliens.dms.common.annotation.UseCase
 import team.aliens.dms.common.spi.SecurityPort
+import team.aliens.dms.domain.notice.service.CommandNoticeService
 import team.aliens.dms.domain.vote.dto.request.CreateVoteTopicRequest
 import team.aliens.dms.domain.vote.exception.NotValidPeriodException
 import team.aliens.dms.domain.vote.model.VotingTopic
-import team.aliens.dms.domain.vote.service.CommendVotingTopicService
-import team.aliens.dms.domain.vote.service.VoteResultNoticeSchedulerService
+import team.aliens.dms.domain.vote.service.VoteService
 
 @UseCase
 class CreateVotingTopicUseCase(
-    private val commendVotingTopicService: CommendVotingTopicService,
-    private val voteResultNoticeSchedulerService: VoteResultNoticeSchedulerService,
-    private val securityPort: SecurityPort
+    private val voteService: VoteService,
+    private val securityPort: SecurityPort,
+    private val noticeService: CommandNoticeService
 ) {
 
     fun execute(request: CreateVoteTopicRequest) {
@@ -23,7 +23,7 @@ class CreateVotingTopicUseCase(
 
         val managerId = securityPort.getCurrentUserId()
 
-        val savedVotingTopicId = commendVotingTopicService.saveVotingTopic(
+        val savedVotingTopicId = voteService.saveVotingTopic(
             VotingTopic(
                 managerId = managerId,
                 topicName = request.topicName,
@@ -34,6 +34,6 @@ class CreateVotingTopicUseCase(
             )
         )
 
-        voteResultNoticeSchedulerService.scheduleVoteResultNotice(savedVotingTopicId, request.endTime)
+        noticeService.scheduleVoteResultNotice(savedVotingTopicId, request.endTime)
     }
 }
