+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/usecase/DeleteVotingTopicUseCase.kt
@@ -2,17 +2,17 @@ package team.aliens.dms.domain.vote.usecase
 
 import team.aliens.dms.common.annotation.UseCase
 import team.aliens.dms.common.spi.TaskSchedulerPort
-import team.aliens.dms.domain.vote.service.CommendVotingTopicService
+import team.aliens.dms.domain.vote.service.VoteService
 import java.util.UUID
 
 @UseCase
 class DeleteVotingTopicUseCase(
-    private val commandVotingTopicService: CommendVotingTopicService,
+    private val voteService: VoteService,
     private val schedulerPort: TaskSchedulerPort
 ) {
 
     fun execute(votingTopicId: UUID) {
-        commandVotingTopicService.deleteVotingTopicById(votingTopicId)
+        voteService.deleteVotingTopicById(votingTopicId)
         schedulerPort.cancelTask(votingTopicId)
     }
 }
