+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/usecase/QueryAllVotingTopicUseCase.kt
@@ -2,12 +2,12 @@ package team.aliens.dms.domain.vote.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
 import team.aliens.dms.domain.vote.dto.reponse.GetAllVotingTopicResponse
-import team.aliens.dms.domain.vote.service.GetVotingTopicService
+import team.aliens.dms.domain.vote.service.VoteService
 
 @ReadOnlyUseCase
 class QueryAllVotingTopicUseCase(
-    val getVotingTopicService: GetVotingTopicService
+    val voteService: VoteService
 ) {
 
-    fun execute() = GetAllVotingTopicResponse.from(getVotingTopicService.getAllVotingTopics())
+    fun execute() = GetAllVotingTopicResponse.from(voteService.getAllVotingTopics())
 }
