+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/usecase/UpdateVotingTopicUseCase.kt
@@ -2,20 +2,18 @@ package team.aliens.dms.domain.vote.usecase
 
 import team.aliens.dms.common.annotation.UseCase
 import team.aliens.dms.common.spi.TaskSchedulerPort
+import team.aliens.dms.domain.notice.service.CommandNoticeService
 import team.aliens.dms.domain.vote.dto.request.UpdateVotingTopicRequest
 import team.aliens.dms.domain.vote.exception.NotValidPeriodException
 import team.aliens.dms.domain.vote.exception.VotingAlreadyEndedException
-import team.aliens.dms.domain.vote.service.CommendVotingTopicService
-import team.aliens.dms.domain.vote.service.GetVotingTopicService
-import team.aliens.dms.domain.vote.service.VoteResultNoticeSchedulerService
+import team.aliens.dms.domain.vote.service.VoteService
 import java.time.LocalDateTime
 
 @UseCase
 class UpdateVotingTopicUseCase(
     private val taskSchedulerPort: TaskSchedulerPort,
-    private val commandVotingTopicService: CommendVotingTopicService,
-    private val getVotingTopicService: GetVotingTopicService,
-    private val voteResultNoticeSchedulerService: VoteResultNoticeSchedulerService,
+    private val voteService: VoteService,
+    private val noticeService: CommandNoticeService
 ) {
 
     fun execute(request: UpdateVotingTopicRequest) {
@@ -23,12 +21,12 @@ class UpdateVotingTopicUseCase(
             throw NotValidPeriodException
         }
 
-        val votingTopic = getVotingTopicService.getVotingTopicById(request.id)
+        val votingTopic = voteService.getVotingTopicById(request.id)
         if (votingTopic.endTime.isBefore(LocalDateTime.now())) {
             throw VotingAlreadyEndedException
         }
 
-        val savedVotingTopicId = commandVotingTopicService.saveVotingTopic(
+        val savedVotingTopicId = voteService.saveVotingTopic(
             votingTopic.copy(
                 topicName = request.topicName,
                 voteType = request.voteType,
@@ -39,6 +37,6 @@ class UpdateVotingTopicUseCase(
         )
 
         taskSchedulerPort.cancelTask(request.id)
-        voteResultNoticeSchedulerService.scheduleVoteResultNotice(savedVotingTopicId, request.endTime)
+        noticeService.scheduleVoteResultNotice(savedVotingTopicId, request.endTime)
     }
 }
